Enhance LocalId parsing and metadata handling
- Added additional parsing states in LocalId for better handling of various metadata tags.
- Improved copy constructor in LocalIdBuilder for better readability and consistency.
- Removed redundant logging statements in LocalIdBuilder's metadata retrieval functions.
- Enhanced error handling for unsupported metadata tags in LocalIdBuilder.
- Updated parsing logic in LocalIdBuilder to handle new states and improve clarity.
- Refactored version parsing in VisVersion to use a map for better maintainability.
- Improved error handling in LocationCharDict for invalid keys.
- Updated Gmod tests to use size_t for occurrence counts for better type safety.

+++ cpp/include/dnv/vista/sdk/Codebooks.h
@@ -27,7 +27,7 @@ namespace dnv::vista::sdk
 			using pointer = const value_type*;
 			using reference = const value_type&;
 
-			Iterator( const std::vector<Codebook>* codebooks, int index );
+			Iterator( const std::vector<Codebook>* codebooks, size_t index );
 
 			reference operator*() const;
 			pointer operator->() const;
@@ -42,7 +42,7 @@ namespace dnv::vista::sdk
 
 		private:
 			const std::vector<Codebook>* m_codebooks;
-			int m_index;
+			size_t m_index;
 			mutable std::optional<value_type> m_current;
 		};
